requirements.txt
Django==1.11.23
django-cors-headers==3.0.2
djangorestframework==3.9.4
Pillow==6.2.2
psycopg2==2.7.3.1
pytz==2019.3
